Added routes for getting data of contacts and partners

+++ backend/app.py
@@ -161,21 +161,22 @@ def edit(id, table):
     if request.method == 'POST':
         # данные из формы
         data = dict(request.form)
-        # файл, загруженный в форму
-        file = request.files['file']
-        if file:
-            if verifyExt(file.filename):
-                # Генерация уникального имени файла
-                unique_filename = f"{uuid.uuid4()}_{file.filename}"
-                # сохранение нового файла
-                file.save(pathlib.Path(UPLOAD_FOLDER, unique_filename))
-                try:
-                    # удаление старого файла из директории
-                    pathlib.Path(UPLOAD_FOLDER, data['url']).unlink()
-                except FileNotFoundError:
-                    print("Не удалось найти файл")
-                # запись нового url в словарь
-                data['url'] = unique_filename
+        if 'file' in request.files:
+            # файл, загруженный в форму
+            file = request.files['file']
+            if file:
+                if verifyExt(file.filename):
+                    # Генерация уникального имени файла
+                    unique_filename = f"{uuid.uuid4()}_{file.filename}"
+                    # сохранение нового файла
+                    file.save(pathlib.Path(UPLOAD_FOLDER, unique_filename))
+                    try:
+                        # удаление старого файла из директории
+                        pathlib.Path(UPLOAD_FOLDER, data['url']).unlink()
+                    except FileNotFoundError:
+                        print("Не удалось найти файл")
+                    # запись нового url в словарь
+                    data['url'] = unique_filename
 
         if table == "news":
             # дата и время изменения записи
@@ -209,17 +210,17 @@ def add_record(table):
     if request.method == 'POST':
         # данные из формы
         data = dict(request.form)
-
-        # файл, загруженный в форму
-        file = request.files['file']
-        if file:
-            if verifyExt(file.filename):
-                # Генерация уникального имени файла
-                unique_filename = f"{uuid.uuid4()}_{file.filename}"
-                # сохранение нового файла
-                file.save(pathlib.Path(UPLOAD_FOLDER, unique_filename))
-                # запись url в словарь
-                data['url'] = file.filename
+        if 'file' in request.files:
+            # файл, загруженный в форму
+            file = request.files['file']
+            if file:
+                if verifyExt(file.filename):
+                    # Генерация уникального имени файла
+                    unique_filename = f"{uuid.uuid4()}_{file.filename}"
+                    # сохранение нового файла
+                    file.save(pathlib.Path(UPLOAD_FOLDER, unique_filename))
+                    # запись url в словарь
+                    data['url'] = file.filename
 
         if table == "news":
             # дата и время изменения записи
